some cleanups, static demo, randomized colors in ambientwall

+++ lib/G35/G35.h
@@ -8,9 +8,10 @@ class G35 {
 		~G35();
 		void setup();
 		void clear(uint8_t r, uint8_t g, uint8_t b, uint8_t i);
-		void set(int x, int y, int r, int g, int b, int i);
-		void setBrightness(byte brightness);
+		void set(uint8_t x, uint8_t y, uint8_t r, uint8_t g, uint8_t b, uint8_t i);
+		void setBrightness(uint8_t brightness);
 		void draw();
+
 	private:
 		void addressMatrix();
 		void sendBulbPacket(uint8_t address, uint8_t pin, uint8_t r, uint8_t g, uint8_t b, uint8_t i);
